backend/requirements.txt
Django==3.0.5
django-colorfield==0.4.3
django-environ==0.7.0
django-filter==21.1
djangorestframework==3.11.0
djoser==2.1.0
drf-extra-fields==3.1.1
Pillow==8.4.1
psycopg2-binary==2.9.1
gunicorn==20.1.0